add unassigned trainee method

+++ app/Http/Resources/OrderItemResource.php
@@ -128,7 +128,7 @@ class OrderItemResource extends JsonResource
                 'sku' => $this->variant->sku,
                 'price' => (float) $this->variant->price,
                 'stock_quantity' => $this->variant->stock_quantity,
-                'is_active' => (bool) $this->variant->is_active,
+                'is_active' => (int) ((bool) $this->variant->is_active),
                 'attributes_string' => $attributesString,
                 'attributes' => $filteredAttributes,
             ],
