+++ app/Http/Resources/ProductVariantResource.php
@@ -29,7 +29,7 @@ class ProductVariantResource extends JsonResource
             'price' => (float) $this->price,
             'base_price_usd' => $this->base_price_usd ? (float) $this->base_price_usd : null,
             'stock_quantity' => (int) $this->stock_quantity,
-            'is_active' => (bool) $this->is_active,
+            'is_active' => (int) ((bool) $this->is_active),
             'is_in_stock' => $isInStock,
             'attributes' => $this->whenLoaded('attributes', function () {
                 return $this->attributes->map(function ($attr) {
